fix : create components companyprofile
